+++ Binance-Spot-API/Model/Market/AggregateTradeList.cs
@@ -8,9 +8,9 @@ namespace Binance_Spot_API.Model.Market
         [JsonProperty("a")]
         public int TradeId { get; set; }
         [JsonProperty("p")]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         [JsonProperty("q")]
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
         [JsonProperty("f")]
         public int FirstTradeId { get; set; }
         [JsonProperty("l")]
