+++ Binance-Spot-API/Model/Market/SymbolOrderBookTicker.cs
@@ -8,12 +8,12 @@ namespace Binance_Spot_API.Model.Market
         [JsonProperty("symbol")]
         public string Symbol { get; set; }
         [JsonProperty("bidPrice")]
-        public decimal BidPrice { get; set; }
+        public double BidPrice { get; set; }
         [JsonProperty("bidQty")]
-        public decimal BidQuantity { get; set; }
+        public double BidQuantity { get; set; }
         [JsonProperty("askPrice")]
-        public decimal AskPrice { get; set; }
+        public double AskPrice { get; set; }
         [JsonProperty("askQty")]
-        public decimal AskQuantity { get; set; }
+        public double AskQuantity { get; set; }
     }
 }
