+++ Binance-Spot-API/Model/Order/Fill.cs
@@ -5,11 +5,11 @@ namespace Binance_Spot_API.Model.Order
     public class Fill
     {
         [JsonProperty("price")]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         [JsonProperty("qty")]
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
         [JsonProperty("commission")]
-        public decimal Commission { get; set; }
+        public double Commission { get; set; }
         [JsonProperty("commissionAsset")]
         public string CommissionAsset { get; set; }
